[FIR2IR] Simplify and unify block creation and coercion to Unit
Get rid of special logic for POSTFIX_INCR/DECR. It was very strange
because we were flattening und then unflattening the blocks.
But it also had a strange interaction with coercion to Unit.
After this commit, block conversion logic is as follows:

Single element blocks are flattened uniformly in
convertToIrExpressionOrBlock and convertToIrBlock.
Coercion to Unit for all but the last expression is inserted in
convertToIrBlock.
For while/do-while and finally blocks, we use the newly introduced
coerceToUnitHandlingSpecialBlocks. It handles blocks with a non-null
origin in a special way by coercing the whole block instead of the last
expression.
This is important for some lowerings like for postfix inc/dec.

This eliminates redundant blocks only containing a single nested block.
And for blocks without an origin, it produces IR more similar to K1 with
unit blocks containing a coerced last expression instead of a block with
some type that is coerced to unit as a whole.

#KT-63348